[update] 記事ComponentのMeta周りを改修
