+++ src/app/articles/[id]/page.tsx
@@ -1,6 +1,6 @@
 import { Article } from "@/components/Article";
-import { getFilePath, getIdToPathMap } from "@/lib/article";
-import { ArticleMetaSchema } from "@/schemas/article/meta";
+import { convertMetaFromRaw, getFilePath, getIdToPathMap } from "@/lib/article";
+import { ArticleRawMetaSchema } from "@/schemas/article/meta";
 import fs from "fs";
 import matter from "gray-matter";
 import type { Metadata } from "next";
@@ -34,7 +34,7 @@ export async function generateMetadata(
   const { id } = await props.params;
   const raw = fs.readFileSync(getFilePath(id), "utf-8");
   const { data } = matter(raw);
-  const validatedMeta = parse(ArticleMetaSchema, data);
+  const validatedMeta = parse(ArticleRawMetaSchema, data);
 
   return {
     title: validatedMeta.title,
@@ -53,7 +53,8 @@ export default async function ArticlePage({ params }: { params: Promise<Params> 
   const filePath = getFilePath(id);
   const raw = fs.readFileSync(filePath, "utf-8");
   const { data, content } = matter(raw);
-  const meta = parse(ArticleMetaSchema, data);
+  const rawMeta = parse(ArticleRawMetaSchema, data);
+  const meta = convertMetaFromRaw(rawMeta);
 
   return (
     <Article meta={ meta } content={ content } />
